+++ src/tauri/walk_forward.rs
@@ -65,8 +65,15 @@ pub struct WalkForwardRunRequest {
     /// The fold count `k`, in `2..=12`. `None` defaults to 6. Wider than the
     /// legal `u8` on purpose: an out-of-range value must reach the domain's
     /// `KOutOfRange` refusal (a `BusError` naming `k`), not die in decoding.
+    ///
+    /// `i32` is the WIDEST signed type specta can export: it refuses the
+    /// `BigInt`-style integers outright (`i64`/`u64`/`usize`/`isize`/`i128`/
+    /// `u128` risk precision loss crossing as a TS `number`), so `i64` here does
+    /// not fail at runtime — it fails the bindings EXPORT, which is a generated
+    /// artifact this repo commits and gates. `i32` still carries `-1` and `256`
+    /// to the typed refusal instead of dying in transport decoding.
     #[serde(default)]
-    pub k: Option<i64>,
+    pub k: Option<i32>,
 }
 
 /// What `get_walk_forward_run` is asked for: one persisted walk-forward run
